requirements.txt
langchain
langchain-core
langchain-groq
python-dotenv
boto3
kubernetes
docker
gitpython
openai
